fix: check for file before delete

+++ src/Generators/Scissors.Xaf.CacheWarmup.Generators.Cli/Program.cs
@@ -27,8 +27,11 @@ namespace Scissors.Xaf.CacheWarmup.Generators.Cli
                 {
                     continue;
                 }
-                Console.WriteLine($"Deleting: {file}");
-                File.Delete(file);
+                if(File.Exists(file))
+                {
+                    Console.WriteLine($"Deleting: {file}");
+                    File.Delete(file);
+                }
             }
 
             AppDomain.CurrentDomain.AssemblyResolve += (object sender, ResolveEventArgs e) =>
@@ -37,7 +40,7 @@ namespace Scissors.Xaf.CacheWarmup.Generators.Cli
                 var assemblyName = e.Name.Split(',').First();
                 var aPath = Path.Combine(assemblyDirectory, assemblyName + ".dll");
                 Console.WriteLine($"New AssemblyPath: '{aPath}'");
-                if (File.Exists(aPath))
+                if(File.Exists(aPath))
                 {
                     Console.WriteLine($"New AssemblyPath exists!: '{aPath}'");
                     var loadedAssembly = Assembly.LoadFile(aPath);
@@ -56,12 +59,12 @@ namespace Scissors.Xaf.CacheWarmup.Generators.Cli
             var foundType = finder.FindAttribute(assembly);
             Console.WriteLine($"Found-Type: '{foundType}'");
 
-            if (foundType != null)
+            if(foundType != null)
             {
                 var cacheGenerator = new CacheWarmupGenerator();
 
                 var cacheResult = cacheGenerator.WarmupCache(assembly, foundType.ApplicationType, foundType.FactoryType);
-                if (cacheResult != null)
+                if(cacheResult != null)
                 {
                     CopyFile(cacheResult.DcAssemblyFilePath, assemblyDirectory);
                     CopyFile(cacheResult.ModelAssemblyFilePath, assemblyDirectory);
@@ -75,12 +78,12 @@ namespace Scissors.Xaf.CacheWarmup.Generators.Cli
         private static void CopyFile(string sourceFile, string destFolder)
         {
             Console.WriteLine($"Copy from {sourceFile} to directory: {destFolder}");
-            if (File.Exists(sourceFile))
+            if(File.Exists(sourceFile))
             {
                 Console.WriteLine($"Copy from {sourceFile} to directory: {destFolder} exists.");
                 var newFile = Path.Combine(destFolder, Path.GetFileName(sourceFile));
                 Console.WriteLine($"Should copy from {sourceFile} to destination: {newFile}");
-                if (File.Exists(newFile))
+                if(File.Exists(newFile))
                 {
                     Console.WriteLine($"Destination exists. Deleting: {newFile}");
                     File.Delete(newFile);
